Fixed navigation to Terms & Privacy screens for Settings and Onboarding screens; got rid of common component for the legal screens.
