+++ app/src/main/java/com/tpp/theperiodpurse/ui/education/EducationHomeScreen.kt
@@ -5,7 +5,6 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.grid.*
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.foundation.text.ClickableText
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -14,21 +13,18 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.platform.UriHandler
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.AnnotatedString
-import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.font.FontWeight.Companion.Bold
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.tpp.theperiodpurse.R
-import com.tpp.theperiodpurse.Screen
 import com.tpp.theperiodpurse.data.Product
 import com.tpp.theperiodpurse.data.ProductsList
+import com.tpp.theperiodpurse.ui.legal.TermsAndPrivacyFooter
 
 const val gray = 0xFF6D6E71
 const val teal = 0xFF72C6B7
@@ -323,47 +319,6 @@ fun SocialMedia(uriHandler: UriHandler) {
         )
     }
 }
-
-
-@Composable
-fun TermsAndPrivacyFooter(navController: NavHostController) {
-    Row(
-        modifier = Modifier
-            .wrapContentHeight()
-            .padding(horizontal = 12.dp),
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-            ClickableText(
-                onClick = { navController.navigate(Screen.Terms.name) },
-                style = TextStyle(
-                    textDecoration = TextDecoration.Underline,
-                    color = Color(teal),
-                    textAlign = TextAlign.Center,
-                    fontSize = 14.sp
-                ),
-                text = AnnotatedString("Terms and Conditions")
-            )
-
-            Text(
-                textAlign = TextAlign.Center,
-                fontSize = 14.sp,
-                text = " and "
-            )
-
-            ClickableText(
-                onClick = { navController.navigate(Screen.Privacy.name) },
-                style = TextStyle(
-                    textDecoration = TextDecoration.Underline,
-                    color = Color(teal),
-                    textAlign = TextAlign.Center,
-                    fontSize = 14.sp
-                    ),
-                text = AnnotatedString("Privacy Policy"),
-            )
-        }
-
-    }
 
 
 /**
