+++ app/src/main/java/com/tpp/theperiodpurse/ui/setting/SettingScreenLayout.kt
@@ -33,14 +33,16 @@ import androidx.compose.ui.unit.sp
 import androidx.navigation.compose.rememberNavController
 import com.tpp.theperiodpurse.R
 import com.tpp.theperiodpurse.ui.education.SocialMedia
-import com.tpp.theperiodpurse.ui.education.TermsAndPrivacyFooter
 import com.tpp.theperiodpurse.ui.education.teal
 import androidx.core.content.ContextCompat
+import androidx.navigation.NavHostController
+import com.tpp.theperiodpurse.ui.legal.TermsAndPrivacyFooter
 
 
 @Composable
 fun SettingScreenLayout(
     modifier: Modifier = Modifier,
+    outController: NavHostController,
     onNotificationClicked: () -> Unit,
     onBackUpClicked: () -> Unit,
     onDeleteClicked: () -> Unit,
@@ -139,7 +141,6 @@ fun SettingScreenLayout(
        Divider(modifier = Modifier.padding(start= 10.dp, end = 10.dp))
        Spacer(modifier = Modifier.padding(20.dp))
        val uriHandler = LocalUriHandler.current
-       val navController = rememberNavController()
        Box(modifier = Modifier.align(Alignment.CenterHorizontally)) {
            SocialMedia(uriHandler)
        }
@@ -156,7 +157,7 @@ fun SettingScreenLayout(
        Terms & Conditions, and Privacy Policy
         */
        Box(modifier = Modifier.align(Alignment.CenterHorizontally)) {
-           TermsAndPrivacyFooter(navController)
+           TermsAndPrivacyFooter(outController)
            Spacer(modifier = Modifier.size(80.dp))
        }
    }
@@ -252,6 +253,7 @@ fun NavigateButton(text: String, onClicked: () -> Unit ){
 fun SettingsScreenPreview() {
     val navController = rememberNavController()
     SettingScreenLayout(
+        outController = navController,
         onNotificationClicked = {
             navController.navigate(SettingScreenNavigation.Notification.name)
         },
